pyproject.toml
[project]
name = "taptools-api-mcp"
version = "0.1.0"
description = "TapTools API integration for MCP (Model Context Protocol)"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.23.0",
    "python-dotenv>=0.21.0",
    "pydantic>=1.10.0",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "rich",
    "structlog",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/taptools_api_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=taptools_api_mcp --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.black]
line-length = 88
target-version = ['py310']
